OPENEUROPA-3020: Refactor preprocess and move rendering test into dedicated class.
