+++ oe_theme.theme
@@ -1399,12 +1399,5 @@ function oe_theme_preprocess_paragraph__oe_facts_figures(array &$variables): voi
 function oe_theme_preprocess_paragraph__oe_timeline(array &$variables): void {
   $paragraph = $variables['paragraph'];
   $variables['limit'] = $paragraph->get('field_oe_timeline_expand')->value;
-  $items = $paragraph->get('field_oe_timeline')->getValue();
-  foreach ($items as $key => $item) {
-    $variables['items'][$key] = [
-      'label' => $item['label'],
-      'title' => $item['title'],
-      'body' => $item['body'],
-    ];
-  }
+  $variables['items'] = $paragraph->get('field_oe_timeline')->getValue();
 }
